fix
The most significant changes include the update to the `AppContext::FatalError` function in `appcontext.cpp` where a comment about the exit code of the application was added. In `czi_brick_reader2.cpp`, the `CziBrickReader2::DoBrick` function was updated with a new instance of `BrickOutputInfo` being created. The `czi_helpers.cpp` file saw a large block of code replaced with a for loop that iterates over `vector_mindex_and_rect`. In `dowarp.cpp`, the `DoWarp::WaitUntilDone` and `DoWarp::TryGetHash` functions were updated, with the latter now being a const function and having a comment about handling exceptions more gracefully. Lastly, the `TaskArenaTbb::TaskArenaTbb` function in `taskarena_tbb.cpp` was updated to capture `this` and `task` in the lambda function.

List of changes:

1. `AppContext::FatalError` function in `appcontext.cpp` updated with a comment about the exit code of the application.
2. Comment removed from `CziBrickReader::ReadBrick` function in `czi_brick_reader.cpp`.
3. `CziBrickReader2::DoBrick` function in `czi_brick_reader2.cpp` updated with a comment removed and a new instance of `BrickOutputInfo` created.
4. Large block of code in `czi_helpers.cpp` replaced with a for loop that iterates over `vector_mindex_and_rect`.
5. Comment in `document_info.h` updated to clarify the definition of a brick.
6. Comment in `BrickInPlaneIdentifier` struct in `document_info.h` updated to clarify the precedence of the `s_index`.
7. `DoWarp::WaitUntilDone` function in `dowarp.cpp` updated to make `TryGetHash` a const function.
8. `DoWarp::TryGetHash` function in `dowarp.cpp` updated with a comment about handling exceptions more gracefully.
9. `dowarp.h` file updated to make `TryGetHash` a const function.
10. `TaskArenaTbb::TaskArenaTbb` function in `taskarena_tbb.cpp` updated to capture `this` and `task` in the lambda function.
11. `TaskArenaTbb::SuspendCurrentTask` function in `taskarena_tbb.cpp` updated with a comment about the user-managed activity that processes async requests.
12. `TEST(Czi_Helpers, GetSubblocksForBrickNoMindexTest)` function in `czi_helpers_tests.cpp` updated to remove a comment and use `TileIdentifier::GetForNoMIndexAndNoSceneIndex()` instead of `numeric_limits<int>::min()`.